Refactor leasehold case attributes out and implemented a way to store leasehold actions
